commands: remove command.py, update docs
Move the contents of command.py to the west.commands package
__init__.py. This makes it possible to use autoclass on the
CommandError subclasses with :show-inheritance: in documentation.

Update some other docstrings to make documentation look nicer.

No functional change expected.
